chore(model): reduce dependency of model from secrets
secrets crate brings some heavy-weight dependencies like
vault-rs and other. We can start compiling model earlier
if this dependency is reduced.
