+++ crates/api-model/Cargo.toml
@@ -35,7 +35,6 @@ carbide-network = { path = "../network", features = ["sqlx"] }
 carbide-version = { path = "../version" }
 carbide-health-report = { path = "../health-report" }
 carbide-ipxe-renderer = { path = "../ipxe-renderer" }
-carbide-secrets = { path = "../secrets" }
 carbide-utils = { path = "../utils", features = ["sqlx"] }
 carbide-uuid = { path = "../uuid", features = ["sqlx"] }
 dns-record = { path = "../dns-record" }
@@ -87,7 +86,9 @@ carbide-version = { path = "../version" }
 
 [dev-dependencies]
 carbide-sqlx-testing = { path = "../sqlx-testing", default-features = false }
+carbide-secrets = { path = "../secrets" }
 lazy_static = { workspace = true }
+p256 = { workspace = true }
 prost = { workspace = true }
 toml = { workspace = true }
